\name{OneStep-package}
\alias{OneStep-package}
\alias{OneStep}
\docType{package}
\title{
\packageTitle{OneStep}
}
\description{
\packageDescription{OneStep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OneStep}
\packageIndices{OneStep}
}
\author{
\packageAuthor{OneStep}

Maintainer: \packageMaintainer{OneStep}
}
\references{
L. LeCam (1956). 
\emph{On the asymptotic theory of estimation and testing hypothesis}. 
In: Proceedings of 3rd Berkeley Symposium I, pages 355-368.
}
\keyword{ package }
\seealso{
See \code{\link[fitdistrplus:fitdistrplus]{fitdistrplus}} for classic MLE, MME,...
}
